# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sushi API",
    "description": "Документация",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/order/view": {
      "post": {
        "tags": [
          "order"
        ],
        "summary": "Просмотр заказов",
        "description": "Просмотр заказов",
        "operationId": "orderView",
        "requestBody": {
          "description": "Апи Токен",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "api_token"
                ],
                "properties": {
                  "api_token": {
                    "type": "string",
                    "format": "string",
                    "example": "123"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Возврощается заказы",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "user": {
                      "type": "object",
                      "example": []
                    },
                    "token": {
                      "type": "string",
                      "example": "FKOhXAr6Xhx2e6fMdaKZbTOCxCBwLuJDO3j8fYjRoDG9XoAYKQUSPzayU4BM"
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  }
}